refactor: split simulate.py into focused modules (SRP)
Extract checkpoint detection into checkpoint.py (~170 lines) and
trajectory validation + structure file discovery into trajectory.py
(~140 lines). simulate.py drops from 871 to 592 lines, retaining
only orchestration dispatch logic.

Update __init__.py to re-export find_structure_file from trajectory
module. Update all test imports and mock patch targets.
